+++ packages/api-reference/src/components/Content/ReferenceEndpoint/ExampleRequest.vue
@@ -18,6 +18,7 @@ import { useGlobalStore } from '../../../stores'
 import { useTemplateStore } from '../../../stores/template'
 import type { TransformedOperation } from '../../../types'
 import { Card, CardContent, CardFooter, CardHeader } from '../../Card'
+import TextSelect from './TextSelect.vue'
 
 const props = defineProps<{
   operation: TransformedOperation
@@ -122,38 +123,32 @@ computed(() => {
         <slot name="header" />
       </div>
       <template #actions>
-        <div class="language-select">
-          <span>
-            {{ getTargetTitle(state.selectedClient) }}
-            {{ getClientTitle(state.selectedClient) }}
-          </span>
-          <select
-            :value="JSON.stringify(state.selectedClient)"
-            @input="
-              (event) =>
-                setItem(
-                  'selectedClient',
-                  JSON.parse((event.target as HTMLSelectElement).value),
-                )
-            ">
-            <optgroup
-              v-for="target in availableTargets"
-              :key="target.key"
-              :label="target.title">
-              <option
-                v-for="client in target.clients"
-                :key="client.key"
-                :value="
-                  JSON.stringify({
-                    targetKey: target.key,
-                    clientKey: client.key,
-                  })
-                ">
-                {{ client.title }}
-              </option>
-            </optgroup>
-          </select>
-        </div>
+        <TextSelect
+          :modelValue="JSON.stringify(state.selectedClient)"
+          :options="
+            availableTargets.map((target) => {
+              return {
+                value: target.key,
+                label: target.title,
+                options: target.clients.map((client) => {
+                  return {
+                    value: JSON.stringify({
+                      targetKey: target.key,
+                      clientKey: client.key,
+                    }),
+                    label: client.title,
+                  }
+                }),
+              }
+            })
+          "
+          @update:modelValue="
+            (value) => setItem('selectedClient', JSON.parse(value))
+          ">
+          {{ getTargetTitle(state.selectedClient) }}
+          {{ getClientTitle(state.selectedClient) }}
+        </TextSelect>
+
         <button
           class="copy-button"
           type="button"
@@ -212,55 +207,6 @@ computed(() => {
 .request-method--put {
   color: var(--theme-color-orange, var(--default-theme-color-orange));
 }
-.language-select {
-  position: relative;
-  padding-right: 9px;
-  height: fit-content;
-  padding-left: 12px;
-  border-right: 1px solid
-    var(--theme-border-color, var(--default-theme-border-color));
-}
-.language-select select {
-  border: none;
-  outline: none;
-  cursor: pointer;
-  background: var(--theme-background-3, var(--default-theme-background-3));
-  box-shadow: -2px 0 0 0
-    var(--theme-background-3, var(--default-theme-background-3));
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  opacity: 0;
-  -moz-appearance: none;
-  -webkit-appearance: none;
-  appearance: none;
-}
-.language-select span {
-  font-size: var(--theme-mini, var(--default-theme-mini));
-  color: var(--theme-color-2, var(--default-theme-color-2));
-  font-weight: var(--theme-semibold, var(--default-theme-semibold));
-  white-space: nowrap;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-.language-select:hover span {
-  color: var(--theme-color-1, var(--default-theme-color-1));
-}
-.language-select span:after {
-  content: '';
-  width: 7px;
-  height: 7px;
-  transform: rotate(45deg) translate3d(-2px, -2px, 0);
-  display: block;
-  margin-left: 6px;
-  box-shadow: 1px 1px 0 currentColor;
-}
-.language-select span:hover {
-  background: var(--theme-background-2, var(--default-theme-background-2));
-}
 
 .copy-button {
   appearance: none;
@@ -296,9 +242,6 @@ computed(() => {
   height: 13px;
 }
 
-.scalar-card-header-actions {
-  display: flex;
-}
 .scalar-card-footer {
   display: flex;
   justify-content: flex-end;